feat(Dockerfile): download graalpy's standalone build
Newer versions of Netshot feature newer versions of graalvm, which don't include the `gu` package manager.
Therefore, `graalpy` is not available anymore on the Netshot docker image, and needs to be downloaded separately in order to use additional Python packages.

This commit pins the Graalpy version to 24.2.1, adds the required steps to download & extract it, and renames the folder conveniently for easier automation.